deception-net/requirements.txt
# Deep Learning
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Networking & Protocols
scapy>=2.5.0
asyncssh>=2.13.0
aiohttp>=3.8.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
websockets>=11.0

# Infrastructure
docker>=6.1.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Data & Storage
pandas>=2.0.0
redis>=4.6.0
elasticsearch>=8.9.0

# Monitoring & Logging
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
